pyproject.toml
[project]
name = "training-resources-calculator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Nicholas Stranges", email = "nicholastranges4@hotmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "pyyaml",
]

[build-system]
requires = ["uv_build>=0.9.26,<0.10.0"]
build-backend = "uv_build"
